Use Makepad button to hide/show Downloads Panel
