--- src/shared/icon.rs
@@ -1,84 +1,0 @@
-use makepad_widgets::*;
-
-live_design! {
-    import makepad_widgets::base::*;
-    import makepad_widgets::theme_desktop_dark::*;
-    import makepad_draw::shader::std::*;
-
-    Icon = {{Icon}} {
-        width: Fit,
-        height: Fit,
-
-        icon_walk: {
-            margin: {left: 5.0},
-            width: Fit,
-            height: Fit,
-        }
-
-        draw_icon: {
-            uniform rotation_angle: 0.0,
-            fn clip_and_transform_vertex(self, rect_pos: vec2, rect_size: vec2) -> vec4 {
-                let clipped: vec2 = clamp(
-                    self.geom_pos * rect_size + rect_pos,
-                    self.draw_clip.xy,
-                    self.draw_clip.zw
-                )
-                self.pos = (clipped - rect_pos) / rect_size
-
-                // Calculate the texture coordinates based on the rotation angle
-                let angle_rad = self.rotation_angle * 3.14159265359 / 180.0;
-                let cos_angle = cos(angle_rad);
-                let sin_angle = sin(angle_rad);
-                let rot_matrix = mat2(
-                    cos_angle, -sin_angle,
-                    sin_angle, cos_angle
-                );
-                self.tex_coord1 = mix(
-                    self.icon_t1.xy,
-                    self.icon_t2.xy,
-                    (rot_matrix * (self.pos.xy - vec2(0.5))) + vec2(0.5)
-                );
-
-                return self.camera_projection * (self.camera_view * (self.view_transform * vec4(
-                    clipped.x,
-                    clipped.y,
-                    self.draw_depth + self.draw_zbias,
-                    1.
-                )))
-            }
-        }
-
-        draw_bg: {
-            instance color: #0000,
-            fn pixel(self) -> vec4 {
-                return self.color;
-            }
-        }
-    }
-}
-
-#[derive(Live, LiveHook, Widget)]
-pub struct Icon {
-    #[redraw]
-    #[live]
-    draw_bg: DrawQuad,
-    #[live]
-    draw_icon: DrawIcon,
-    #[live]
-    icon_walk: Walk,
-    #[walk]
-    walk: Walk,
-    #[layout]
-    layout: Layout,
-}
-
-impl Widget for Icon {
-    fn handle_event(&mut self, _cx: &mut Cx, _event: &Event, _scope: &mut Scope) {}
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, _scope: &mut Scope, walk: Walk) -> DrawStep {
-        self.draw_bg.begin(cx, walk, self.layout);
-        self.draw_icon.draw_walk(cx, self.icon_walk);
-        self.draw_bg.end(cx);
-        DrawStep::done()
-    }
-}
