+++ src/shared/mod.rs
@@ -3,7 +3,6 @@ use makepad_widgets::Cx;
 pub mod actions;
 pub mod download_notification_popup;
 pub mod external_link;
-pub mod icon;
 pub mod portal;
 pub mod modal;
 pub mod resource_imports;
@@ -15,7 +14,6 @@ pub fn live_design(cx: &mut Cx) {
     styles::live_design(cx);
     resource_imports::live_design(cx);
     widgets::live_design(cx);
-    icon::live_design(cx);
     portal::live_design(cx);
     modal::live_design(cx);
     external_link::live_design(cx);
